+++ product/service/domain/accounting/MCF.cs
@@ -4,7 +4,7 @@ namespace solidware.financials.service.domain.accounting
 {
     public class MCF : SimpleUnitOfMeasure
     {
-        public override string pretty_print(double amount)
+        public override string pretty_print(decimal amount)
         {
             throw new NotImplementedException();
         }
