+++ product/service/domain/accounting/UnitOfMeasure.cs
@@ -2,8 +2,8 @@ namespace solidware.financials.service.domain.accounting
 {
     public interface UnitOfMeasure
     {
-        double convert(double amount, UnitOfMeasure other);
-        string pretty_print(double amount);
+        decimal convert(decimal amount, UnitOfMeasure other);
+        string pretty_print(decimal amount);
     }
 
 }
